restapi/python-mock-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
requests==2.31.0
